Embeddings now are stored and loaded from one file
Embeddings now are loaded from one file, and when that file does not exist,
the embeddings are generated.

The format now is different since the paths to the embeddings is no longer
necessary, and the src and trg embeddings files are provided with positional
arguments.

Scripts have been created in order to generate this merged embedding file
from a list of paths and viceversa.

Language pair is no longer required since it was only necessary for the
sentence splitting, which now is optional, but the language pair will be
necessary when the sentence splitting have to be performed.

Other changes and fixes.