More verbose, fixes and new features
Minor fixes about the default parameter values and f-strings.

Now, the content of the documents, or Base64 values, will be
clipped in order to avoid VERY long lines.

New limit when genereting embeddings using the number of lines
and not only the size in bytes of the content. This new limit is
useful since the quantity of lines is what can be directly
translated to memory, disk and processment resources, so there will
be more control over the resources (both limits are applied, but
they can be disabled as well).